requirements.txt
hydra-core==1.1.2
numpy==1.21.6
pandas==1.3.5
pytorch-lightning==1.6.3
streamlit==1.9.0
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.2
matplotlib==3.5.2
sox==1.4.1
PySoundFile==0.9.0.post1
bpe==1.0